Define new class to represent editor document
A new class was created that will hold information about the
document the user creates in the editor. This class defines
two methods that should output the user's content in two different
formats: as HTML, and as an array of objects that represent the
content of the document. Currently, these two methods both return
undefined.

Two unit tests were written to test the functionality of these two
methods.