Define models and controller class
A new controller class was defined that will handle all actions
performed by the user to edit the document. Several model classes
were also defined to represent the document and handle rendering
of the document.

I feel like, during the first week that I was working on this
project, I was just coding aimlessly. I had no idea how I was
going to structure the code or how to implement any of the
functionality that the editor would need to perform.
I've finally come up with a plan to help structure everything.
The editor will be implemented using a model-view-controller
architecture. A model class will be used to describe the
structure of the document and the HTML elements that it is
comprised of. A controller class will be used to execute the
actions performed by the user. After each action is performed,
the document model will then be updated and used to render the
HTML on the page.