Define methods of controlled to hand key input
Several new methods were defined for the editor controller class.
These methods handle keyup events initiated by the user. If the
user types a character, that character is added to the document
model object. Upon every key stroke, whether it comes from a
character key or not, the current position of the caret is recorded
by the controller.