Change methods for EditorController class
The functionality of the pressKey method of the EditorController
class was updated. It now creates a new Command object, or updates
an existing Command object, every time a user types a key. This
will be important later for keeping track of a user's command
history.

The functionality of the flush method still needs to be defined
so that each new Command is actually applied to the document
object.