Define more unit tests for the EditorController
Some more tests were defined to run against the EditorController
class. These tests ensure that the controller accepts text being
typed by the user and applies the necessary changes to the
document.

A few tests were written to run against the EditorModel class, but
these tests currently fail. The applyCommand method of the
EditorModel class needs to be updated to actually apply user
commands to the document.