Define part of the applyCommand method
Part of the applyCommand method of the EditorModel class was
defined. Now, this class passes one of the unit tests. The method
still needs to be modified to get the remaining tests to pass.
This will ensure that any text the user types is included in the
object that represents the document.